requirements.txt
flipside==2.0.8
idna==3.4
pydantic==1.10.9
pytest==7.4.0
typing_extensions==4.6.3
urllib3==1.26.16
python-dotenv==1.0.0